Check BST.
In this method we test wether the given tree is a bst or not. using the idea that For all root of each subtree has to be greater that its left child and smaller than its right child. i.e. Recursive solution.
Here the idea is to find left max value for each root and also the right min value for each node. Collectively we have left max,right max for our root of the tree, we do then comparisions. Finally we test it for each subtree of the given tree,